% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_tuv_dir}
\alias{clean_tuv_dir}
\title{Delete the directory containing the TUV files}
\usage{
clean_tuv_dir(tuv_dir = tuv_data_dir())
}
\arguments{
\item{tuv_dir}{The directory containing the TUV files}
}
\description{
Delete the directory containing the TUV files
}
