% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Get elevation for a lon/lat pair}
\usage{
get_elevation(lon, lat)
}
\arguments{
\item{lon}{Longitude: number between -140 and -53}

\item{lat}{Latitude: number 41 and 84}
}
\value{
Elevation, in m
}
\description{
Look up elevation for a point location using Natural Resource Canada's
Elevation API: https://natural-resources.canada.ca/science-and-data/science-and-research/earth-sciences/geography/topographic-information/web-services/elevation-api/17328,
and if outside of Canada, using the USGS Elevation Point Query Service:
https://epqs.nationalmap.gov/v1/docs
}
\examples{
get_elevation(-115, 53)
}
