% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-tuv.R
\name{get_tuv_results}
\alias{get_tuv_results}
\title{Retrieve the results of a TUV run}
\usage{
get_tuv_results(file = "out_irrad_y", tuv_dir = tuv_data_dir())
}
\arguments{
\item{file}{one of "out_irrad_y", "out_aflux_y", "out_irrad_ave",
"out_aflux_ave", "out_irrad_atm", "out_aflux_atm". Default is \code{"out_irrad_y"},
for input into \code{\link[=p_abs]{p_abs()}}.}

\item{tuv_dir}{the directory where the compiled TUV executable is located}
}
\value{
A data.frame with the results of the TUV run. It contains columns
specifying wavelength and calculated Kd(lambda), and column for each timestamp
containing the irradiance at that time and wavelength, in W m-2 nm-1.
}
\description{
Retrieve the results of a TUV run
}
