% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kd-calc.R
\name{kd_305}
\alias{kd_305}
\title{Calculate Kd at 305 nm for a given Dissolved Organic Carbon (DOC) concentration.}
\usage{
kd_305(DOC)
}
\arguments{
\item{DOC}{DOC in g/m^3}
}
\value{
A numeric vector representin Kd at 305 nm
}
\description{
Calculate Kd at 305 nm for a given Dissolved Organic Carbon (DOC) concentration.
}
\examples{
kd_305(5)
}
