% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kd-calc.R
\name{kd_lambda}
\alias{kd_lambda}
\title{Calculate Kd at a given wavelength and DOC concentration.}
\usage{
kd_lambda(DOC, lambda)
}
\arguments{
\item{DOC}{DOC in g/m^3}

\item{lambda}{wavelength in nm}
}
\value{
A numeric vector representin Kd at a given wavelength
}
\description{
Calculate Kd at a given wavelength and DOC concentration.
}
\examples{
kd_lambda(10, 400)
}
