% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kd-calc.R
\name{kd_marine}
\alias{kd_marine}
\title{Calculate Kd at a given wavelength in marine waters.}
\usage{
kd_marine(wavelength)
}
\arguments{
\item{wavelength}{lambda wavelength in nm}
}
\value{
A numeric vector representing Kd at a given wavelength in marine waters
}
\description{
Calculate Kd at a given wavelength in marine waters.
}
\examples{
kd_marine(400)
}
