% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molar_absorption}
\alias{molar_absorption}
\title{Molar absorption coefficients for multiple PAHs at different wavelengths.
Taken from Table S5-a. Molar absorption spectra epsilon(lambda) at different
wavelength lambda (nm) in:}
\format{
\subsection{\code{molar_absorption} A data frame with 132 rows and 11 columns:}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{One column for `wavelength` (nm), and a column for each PAH, with values of
molar absorption values at each wavelength. Units are (L/mol/cm).
}\if{html}{\out{</div>}}
}
}
\source{
< https://doi.org/10.1002/etc.3601>
}
\usage{
molar_absorption
}
\description{
Marzooghi, S., Finch, B. E., Stubblefield, W. A., Dmitrenko, O., Neal, S. L.,
& Di Toro, D. M. (2017). Phototoxic target lipid model of single polycyclic
aromatic hydrocarbons. Environmental toxicology and chemistry, 36(4),
926-937.
}
\keyword{datasets}
