% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlm.R
\name{narcotic_benchmark}
\alias{narcotic_benchmark}
\title{Calculate the narcotic benchmark (acute) concentration for a PAH or HAC using
the Target Lipid Model (TLM)}
\usage{
narcotic_benchmark(
  chemical,
  slope = -0.922,
  HC5 = 9.7,
  dc_pah = -0.42,
  dc_hac = -0.467
)
}
\arguments{
\item{chemical}{The chemical (a HAC or PAH) of interest}

\item{slope}{The slope in Equation 2 in Tillmanns et al 2024. The default
value is -0.922. It is not recommended to adjust this without good
justification.}

\item{HC5}{The 5th percentile of the SSD of critical body burdens predicted
to be hazardous for no more than 5\% of the species. Default value is 9.7
umol/g, from Equation 2 in Tillmanns et al 2024. It is not recommended to
adjust this without good justification.}

\item{dc_pah}{Chemical class correction (Δc) for PAHs, as reported in
Tillmanns et al 2024. The default value is -0.420.}

\item{dc_hac}{Chemical class correction (Δc) for HACs, as reported in
Tillmanns et al 2024. The default value is -0.467.}
}
\value{
the narcotic benchmark value of the PAH in ug/L.
}
\description{
This calculates the acute water quality benchmark using the equation and
default values from Tillmanns et al 2024.
}
\examples{
narcotic_benchmark("anthracene")
}
\references{
Tillmanns, A. R., McGrath, J. A., & Di Toro, D. M. (2024). International
Water Quality Guidelines for Polycyclic Aromatic Hydrocarbons: Advances to
Improve Jurisdictional Uptake of Guidelines Derived Using The Target Lipid
Model. Environmental Toxicology and Chemistry, 43(4), 686-700.
}
