% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlm.R
\name{narcotic_cwqg}
\alias{narcotic_cwqg}
\title{Calculate the narcotic guideline (chronic) concentration for a PAH or HAC using
the Target Lipid Model (TLM)}
\usage{
narcotic_cwqg(chemical)
}
\arguments{
\item{chemical}{The chemical (a HAC or PAH) of interest}
}
\value{
the narcotic chronic water quality guideline value of the PAH in ug/L.
}
\description{
This calculates the narcotic chronic water quality guideline using the equation and
default values from Tillmanns et al 2024.
}
\details{
The values used in the calculation are:
\itemize{
\item \strong{slope} The slope in Equation 2 in Tillmanns et al 2024. The default
value is -0.951.
\item \strong{HC5} The 5th percentile of the SSD of critical body burdens predicted
to be hazardous for no more than 5\% of the species. Default value is 3.14
umol/g, from Equation 2 in Tillmanns et al 2024.
\item \strong{dc_pah} Chemical class correction (Δc) for PAHs, as reported in
Tillmanns et al 2024. The default value is -0.420.
\item \strong{dc_hac} Chemical class correction (Δc) for HACs, as reported in
Tillmanns et al 2024. The default value is -0.467.
}
}
\examples{
narcotic_benchmark("anthracene")
}
\references{
Tillmanns, A. R., McGrath, J. A., & Di Toro, D. M. (2024). International
Water Quality Guidelines for Polycyclic Aromatic Hydrocarbons: Advances to
Improve Jurisdictional Uptake of Guidelines Derived Using The Target Lipid
Model. Environmental Toxicology and Chemistry, 43(4), 686-700.
}
