% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlm.R
\name{nlc50_lookup}
\alias{nlc50_lookup}
\title{Look up the NLC50 value for a PAH.}
\usage{
nlc50_lookup(pah)
}
\arguments{
\item{pah}{The PAH of interest}
}
\value{
NLC50 value, in ug/L
}
\description{
Look up the NLC50 value for a PAH.
}
\examples{
nlc50_lookup("anthracene")
}
