% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlm.R
\name{p_abs_single}
\alias{p_abs_single}
\title{Calculate the light absorption of a PAH from a single exposure experiment}
\usage{
p_abs_single(exposure, pah, time_multiplier = 1)
}
\arguments{
\item{exposure}{two-column data.frame of exposure results. The first column
must contain the wavelenghts and be called \code{wl}, the second column
must contain the irradiance values at each wavelength in
units of (mol photon cm3) / (uW s nm L)}

\item{pah}{name of PAH to calculate light absorption for}

\item{time_multiplier}{multiplier to get the total exposure time. I.e., if
the exposure was one second, and you need a 48h exposure, the
multiplier would be 3600 * 48}
}
\value{
The value of \code{Pabs} for the exposure results.
}
\description{
Calculate the light absorption of a PAH from a single exposure experiment
}
