% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pahwq-package.R
\docType{package}
\name{pahwq-package}
\alias{pahwq}
\alias{pahwq-package}
\title{pahwq: Calculate site-specific water quality guidelines for PAHs}
\description{
Calculate water quality guidlines for polycyclic aromatic hydrocarbons (PAHs) for a specific location, taking into account the phototoxicity of specific PAHs at a given depth. Light absorption at depth is calculated using the TUV model ('https://github.com/NCAR/TUV').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/pahwq}
  \item \url{https://bcgov.github.io/pahwq/}
  \item Report bugs at \url{https://github.com/bcgov/pahwq/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
  \item National Center for Atmospheric Research (TUV model) [copyright holder]
  \item Canadian Council of Ministers of the Environment (Modifications of TUV model) [copyright holder]
}

}
\keyword{internal}
