% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plc50-multi.R
\name{pb_multi}
\alias{pb_multi}
\title{Calculate the narcotic and phototoxic benchmarks for a set of PAHs}
\usage{
pb_multi(tuv_results, pahs, time_multiplier = 2, ...)
}
\arguments{
\item{tuv_results}{data.frame of TUV results}

\item{pahs}{names of PAHs for which to calculate narcotic benchmark, Pabs,
and phototoxic benchmark}

\item{time_multiplier}{multiplier to get the total exposure time. I.e., if
the tuv_results contains 24 hours of data, and you need a 48h exposure, the
multiplier would be 2. (this is the default)}

\item{...}{arguments passed on to \code{\link[=narcotic_benchmark]{narcotic_benchmark()}}}
}
\value{
a data.frame of narcotic benchmark, Pabs, and phototoxic benchmark
for the given PAHs and TUV results
}
\description{
Given the results of a TUV model run (via \code{\link[=tuv]{tuv()}} or \code{\link[=run_tuv]{run_tuv()}}),
get the narcotic benchmark, Pabs, and phototoxic benchmark for a set of PAHs
}
\examples{
tuv_res <- tuv(
  depth_m = 0.25,
  lat = 49.601632,
  lon = -119.605862,
  DOC = 5,
  date = "2023-06-21"
)

pb_multi(tuv_res, c("Anthracene", "fluorene", "pyrene"))
}
