% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlm.R
\name{phototoxic_benchmark}
\alias{phototoxic_benchmark}
\title{Calculate the phototoxic benchmark for a given P~abs~ and PAH chemical using
the PTLM}
\usage{
phototoxic_benchmark(x, pah = NULL, narc_bench = NULL, time_multiplier)
}
\arguments{
\item{x}{light absorption, calculated from \code{\link[=p_abs]{p_abs()}}, or a \code{tuv_results}
data.frame from \code{\link[=tuv]{tuv()}} or \code{\link[=get_tuv_results]{get_tuv_results()}}.}

\item{pah}{The PAH of interest, which is used to calculate the narcotic
benchmark value.}

\item{narc_bench}{(optional) the narcotic toxicity (i.e., in the
absence of light) of the PAH in ug/L. If supplied, takes precedence over
the PAH lookup.}

\item{time_multiplier}{If x is a \code{tuv_results} data frame, this is the
multiplier to get the total exposure time. I.e., if the tuv_results
contains 24 hours of data, and you need a 48h exposure, the
multiplier would be 2 (this is the default). Ignored if \code{x} is a numeric
value of light absorption.}
}
\value{
the phototoxic benchmark value of the PAH in ug/L.
}
\description{
The phototoxic benchmark is the acute benchmark concentration of a phototoxic
PAH based on its narcotic toxicity (narcotic benchmark) and calculations of
site-specific or field-level light absorption.
}
\details{
You can either supply a specific PAH, so the narcotic benchmark can be
calculated for that chemical, or supply a narcotic benchmark value directly.
}
\examples{
phototoxic_benchmark(590, pah = "Benzo[a]pyrene")
phototoxic_benchmark(590, narc_bench = 450)
}
\references{
Marzooghi, S., Finch, B.E., Stubblefield, W.A., Dmitrenko, O., Neal, S.L. and
Di Toro, D.M. (2017), Phototoxic target lipid model of single polycyclic
aromatic hydrocarbons. Environ Toxicol Chem, 36: 926-937.
https://doi.org/10.1002/etc.3601
}
