% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlm.R
\name{plc50}
\alias{plc50}
\title{Calculate the PLC50 for a given P~abs~ and PAH chemical using the PTLM}
\usage{
plc50(x, pah = NULL, NLC50 = NULL)
}
\arguments{
\item{pah}{The PAH of interest, which is used to look up the NLC50.}

\item{NLC50}{(optional) the narcotic toxicity (i.e., in the absence of light)
of the PAH in ug/L. If supplied, takes precedence over the PAH lookup.}

\item{p_abs}{light absorption, calculated from \code{p_abs()}}
}
\value{
the PLC50 of the PAH.
}
\description{
PLC50 is the LC50 of a phototoxic PAH based on calculations of site-specific
or field-level light absorption.
}
\details{
You can either supply a specific PAH, so the NLC50 can be calculated for
that chemical, or supply a NLC50 value directly.
}
\examples{
plc50(590, pah = "Benzo[a]pyrene")
plc50(590, NLC50 = 450)
}
\references{
Marzooghi, S., Finch, B.E., Stubblefield, W.A., Dmitrenko, O., Neal, S.L. and
Di Toro, D.M. (2017), Phototoxic target lipid model of single polycyclic
aromatic hydrocarbons. Environ Toxicol Chem, 36: 926-937.
https://doi.org/10.1002/etc.3601
}
