% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plc50-multi.R
\name{plc50_multi}
\alias{plc50_multi}
\title{Calculate the NLC50 and PLC50 for a set of PAHs}
\usage{
plc50_multi(tuv_results, pahs)
}
\arguments{
\item{tuv_results}{data.frame of TUV results}

\item{pahs}{names of PAH to calculate light absorption for}
}
\value{
a data.frame of NLC50, Pabs, and PLC50 for the given PAHs and TUV
results
}
\description{
Given the results of a TUV model run (via \code{\link[=tuv]{tuv()}} or \code{\link[=run_tuv]{run_tuv()}}),
get the NLC50, Pabs, and PLC50 for a set of PAHs
}
\examples{
tuv_res <- tuv(
  depth_m = 0.25,
  lat = 49.601632,
  lon = -119.605862,
  DOC = 5,
  date = "2023-06-21"
)

plc50_multi(tuv_res, c("Anthracene", "fluorene", "pyrene"))
}
