% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlm.R
\name{plc_50}
\alias{plc_50}
\title{Calculate the PLC50 for a given Pabs and NLC50.}
\usage{
plc_50(p_abs, NLC50)
}
\arguments{
\item{p_abs}{light absorption, calculated from \code{p_abs()}}

\item{NLC50}{the narcotic toxicity (i.e., in the absence of light) of the PAH.}
}
\value{
the PLC50 of the PAH.
}
\description{
PLC50 is the LC50 of a phototoxic PAH based on calculations of site-specific
or field-level light absorption.
}
\examples{
plc_50(590, 450)
}
