% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-kd-depth.R
\name{plot_sens_kd_depth}
\alias{plot_sens_kd_depth}
\title{Make a heatmap of the sensitivity analysis performed by \code{\link[=sens_kd_depth]{sens_kd_depth()}}}
\usage{
plot_sens_kd_depth(x)
}
\arguments{
\item{x}{A data.frame, the output of \code{\link[=sens_kd_depth]{sens_kd_depth()}}}
}
\value{
a ggplot2 object
}
\description{
Make a heatmap of the sensitivity analysis performed by \code{\link[=sens_kd_depth]{sens_kd_depth()}}
}
\examples{
out <- sens_kd_depth(
  "Anthracene",
  lat = 52,
  lon = -113,
  Kd_ref = 40:45,
  depth_m = c(0.25, 0.5, 1),
  date = c("2023-07-01", "2023-08-01")
)
plot_sens_kd_depth(out)
}
