% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-tuv.R
\name{run_tuv}
\alias{run_tuv}
\title{Call the TUV program}
\usage{
run_tuv(tuv_dir = tuv_data_dir(), quiet = FALSE)
}
\arguments{
\item{tuv_dir}{the directory where the compiled TUV executable is located}

\item{quiet}{Should the progress of the TUV program be printed to the console?}
}
\description{
You must set tuv parameters by calling \code{\link[=set_tuv_aq_params]{set_tuv_aq_params()}} before calling
\code{run_tuv()}
}
