% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-kd-depth.R
\name{sens_kd_depth}
\alias{sens_kd_depth}
\title{Sensitivity analysis for DOC, depth, and time of year}
\usage{
sens_kd_depth(
  pah = NULL,
  lat = NULL,
  lon = NULL,
  elev_m = NULL,
  DOC = NULL,
  Kd_ref = NULL,
  depth_m = NULL,
  date = NULL,
  time_multiplier = 2,
  ...
)
}
\arguments{
\item{pah}{The PAH of interest, which is used to look up the NLC50.}

\item{lat}{latitude of the site, decimal degrees. Required.}

\item{lon}{longitude of the site, decimal degrees. Required.}

\item{elev_m}{elevation of the site above sea level, in metres. If \code{NULL}
(default) it is looked up using a digital elevation model using lat and
lon.}

\item{DOC}{dissolved organic carbon concentration, in mg/L. Ignored if
\code{Kd_ref} is set directly.}

\item{Kd_ref}{Light attenuation coefficient at reference wavelength. Can be
set directly, or calculated from \code{DOC}.}

\item{depth_m}{depth at which to calculate the light attenuation coefficient.
Required.}

\item{date}{date of the calculation, as \code{Date} object, or a character in a
standard format that can be converted to a \code{Date} object (e.g.,
"YYYY-MM-DD"). Required.}

\item{time_multiplier}{If x is a \code{tuv_results} data frame, this is the
multiplier to get the total exposure time. I.e., if the tuv_results
contains 24 hours of data, and you need a 48h exposure, the
multiplier would be 2 (this is the default). Ignored if \code{x} is a numeric
value of light absorption.}

\item{...}{other parameters passed on to the tuv model. See \code{\link[=tuv]{tuv()}}}
}
\value{
A data.frame of all of the input parameters, plus a list column
containing TUV results and NLC50, Pabs, and PLC50 values
}
\description{
Enter a single location and PAH, and a range of values for \code{DOC} or \code{Kd_ref},
water depth, and dates and get back a data.frame of NLC50, Pabs, and PLC50
values.
}
\details{
You can add other variables beyond those listed explicitly, but if there are
too many combinations it will create many runs of the TUV model, which can
take a long time. Explicit variable checking is only performed on \code{pah},
\code{lat}, \code{lon}, \code{elev_m}, \code{DOC}, \code{Kd_ref}, \code{depth}, and \code{months}. Passing
invalid values of other parameters may cause cryptic errors or unexpected
results.
Note that combinations of many \code{DOC}, \code{Kd_ref}, \code{depth_m}, and \code{date} values
will result in many runs of the TUV model and thus take a long time.
}
\examples{
sens_kd_depth(
  "Anthracene",
  lat = 52,
  lon = -113,
  Kd_ref = 40:45,
  depth_m = c(0.25, 0.5, 1),
  date = c("2023-07-01", "2023-08-01")
)
}
