% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-tuv.R
\name{setup_tuv_options}
\alias{setup_tuv_options}
\title{Title}
\usage{
setup_tuv_options(
  depth_m = NULL,
  lat = NULL,
  lon = NULL,
  elev_km = NULL,
  date = NULL,
  DOC = NULL,
  tzone = 0L,
  tstart = 0,
  tstop = 23,
  tsteps = 24L,
  wvl_start = 279.5,
  wvl_end = 400.5,
  wvl_steps = 121L,
  ...,
  tuv_dir = getOption("tuv_dir", default = NULL),
  write = TRUE
)
}
\arguments{
\item{depth_m}{depth at which to calculate the light attenuation coefficient.
Required.}

\item{lat}{latitude of the site, decimal degrees. Required.}

\item{lon}{longitude of the site, decimal degrees. Required.}

\item{elev_km}{elevation of the site above sea level, in kilometres. Required.}

\item{date}{date of the calculation, as \code{Date} object, or a character in a
standard format that can be converted to a \code{Date} object (e.g.,
"YYYY-MM-DD"). Required.}

\item{DOC}{dissolved organic carbon concentration, in mg/L. Required.}

\item{tzone}{timezone offset from UTC, in hours. Default \code{0}.}

\item{tstart}{start time of the calculation, in hours. Default \code{0}.}

\item{tstop}{stop time of the calculation, in hours. Default \code{23}.}

\item{tsteps}{number of time steps to calculate. Default \code{24}.}

\item{wvl_start}{start wavelength of the calculation, in nm. Default \code{279.5}.}

\item{wvl_end}{end wavelength of the calculation, in nm. Default \code{400.5}.}

\item{wvl_steps}{number of wavelength steps to calculate. Default \code{121}.}

\item{...}{other options passed on to the TUV model. See \code{\link[=inp_aq_defaults]{inp_aq_defaults()}}}

\item{tuv_dir}{the directory where the compiled TUV executable is located}

\item{write}{should the options be written to \code{inp_aq} in the TUV directory? Default \code{TRUE}.}
}
\description{
Title
}
