% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-tuv.R
\name{tuv_aq_defaults}
\alias{tuv_aq_defaults}
\title{Get a list of TUV inputs and their default values}
\usage{
tuv_aq_defaults()
}
\value{
a list of TUV inputs and their default values
}
\description{
Inputs that don't have a default value and thus are required to be specified
in \code{set_tup_options()} are shown as an empty vector of the required data type.
}
