% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-tuv.R
\name{tuv_run_params}
\alias{tuv_run_params}
\title{Show the input parameters used for a TUV model run}
\usage{
tuv_run_params(x)
}
\arguments{
\item{x}{tuv results, a result of running \code{\link[=run_tuv]{run_tuv()}}}
}
\value{
a named character vector of the inputs to the TUV model
}
\description{
Show the input parameters used for a TUV model run
}
