% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-tuv.R
\name{get_tuv_results}
\alias{get_tuv_results}
\title{Retrieve results of TUV run}
\usage{
get_tuv_results(
  file = "out_irrad_y",
  tuv_dir = getOption("tuv_dir", default = NULL)
)
}
\arguments{
\item{file}{one of "out_irrad_y", "out_aflux_y", "out_irrad_ave",
"out_aflux_ave", "out_irrad_atm", "out_aflux_atm"}

\item{tuv_dir}{the directory where the compiled TUV executable is located}
}
\value{
A data.frame with the results of the TUV run
}
\description{
Retrieve results of TUV run
}
