% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kd-calc.R
\name{kd_lambda}
\alias{kd_lambda}
\title{Calculate Kd at a given wavelength and DOC concentration.}
\usage{
kd_lambda(DOC, wavelength)
}
\arguments{
\item{DOC}{DOC in g/m^3}

\item{wavelength}{lambda wavelength in nm}
}
\value{
A numeric vector representing Kd at a given wavelength
}
\description{
Note this function is not used inside the package as the same calculation is
done by the Fortran TUV model. It is present here for demonstration purposes.
}
\examples{
kd_lambda(10, 400)
}
