% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_tuv_dir}
\alias{list_tuv_dir}
\title{List files in the directory containing the TUV files}
\usage{
list_tuv_dir(tuv_dir = tuv_data_dir())
}
\arguments{
\item{tuv_dir}{The directory containing the TUV files}
}
\description{
List files in the directory containing the TUV files
}
