% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlm.R
\name{nlc50}
\alias{nlc50}
\title{Calculate the NLC50 value for a PAH using the Target Lipid Model (TLM)}
\usage{
nlc50(pah, slope = -0.94, HC5 = 9.3)
}
\arguments{
\item{pah}{The PAH of interest}

\item{slope}{Slope in Equation 1 in McGrath et al. 2018. Default \code{-0.94}.}

\item{HC5}{The critical target lipid body burden above which 95\% of species
should be protected in μmol/g octanol. Default \code{9.3}.}
}
\value{
NLC50 value, in ug/L
}
\description{
This uses the equation and default values from McGrath et al. 2018.
}
\examples{
nlc50("anthracene")
}
\references{
McGrath, J.A., Fanelli, C.J., Di Toro, D.M., Parkerton, T.F., Redman, A.D.,
Paumen, M.L., Comber, M., Eadsforth, C.V. and den Haan, K. (2018),
Re-evaluation of target lipid model–derived HC5 predictions for hydrocarbons.
Environ Toxicol Chem, 37: 1579-1593. https://doi.org/10.1002/etc.4100
}
