% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-tuv.R
\name{set_tuv_aq_params}
\alias{set_tuv_aq_params}
\title{Set required and optional parameters for TUV}
\usage{
set_tuv_aq_params(
  depth_m = NULL,
  lat = NULL,
  lon = NULL,
  elev_km = NULL,
  date = NULL,
  Kd_ref = NULL,
  Kd_wvl = NULL,
  DOC = NULL,
  tzone = 0L,
  tstart = 0,
  tstop = 23,
  tsteps = 24L,
  wvl_start = 280,
  wvl_end = 420,
  wvl_steps = wvl_end - wvl_start + 1,
  o3_tc = NULL,
  tauaer = NULL,
  ...,
  write = TRUE,
  tuv_dir = tuv_data_dir()
)
}
\arguments{
\item{depth_m}{depth at which to calculate the light attenuation coefficient.
Required.}

\item{lat}{latitude of the site, decimal degrees. Required.}

\item{lon}{longitude of the site, decimal degrees. Required.}

\item{elev_km}{elevation of the site above sea level, in kilometres.
Required.}

\item{date}{date of the calculation, as \code{Date} object, or a character in a
standard format that can be converted to a \code{Date} object (e.g.,
"YYYY-MM-DD"). Required.}

\item{Kd_ref}{Light attenuation coefficient at reference wavelength. Can be
set directly, or calculated from \code{DOC}.}

\item{Kd_wvl}{The reference wavelength at which \code{Kd_ref} was obtained.
Default \code{305}. Only used if \code{Kd_ref} is set.}

\item{DOC}{dissolved organic carbon concentration, in mg/L. Ignored if
\code{Kd_ref} is set directly.}

\item{tzone}{timezone offset from UTC, in hours. Default \code{0}.}

\item{tstart}{start time of the calculation, in hours. Default \code{0}.}

\item{tstop}{stop time of the calculation, in hours. Default \code{23}.}

\item{tsteps}{number of time steps to calculate. Default \code{24}.}

\item{wvl_start}{start wavelength of the calculation, in nm. Default \code{280}.}

\item{wvl_end}{end wavelength of the calculation, in nm. Default \code{400}.}

\item{wvl_steps}{number of wavelength steps to calculate. Default 1 step per
nm from \code{wvl_start} and \code{wvl_end}, inclusive.}

\item{o3_tc}{The ozone column, in Dobson Units. If \code{NULL}, it is looked up
based on latitude and month, based on historic climatology. If there is no
historic value for the given month and location, a default value of 300 is
used. You can force the use of this default by setting the value of this
parameter to \verb{"default}.}

\item{tauaer}{The aerosol optical depth (tau) at 550 nm. If \code{NULL}, it is
looked up based on latitude, longitude, and month, based on historic
climatology. If there is no historic value for the given month and
location, a default value of 0.235 is used. You can force the use of this
default by setting the value of this parameter to \verb{"default}.}

\item{...}{other options passed on to the TUV model. See \code{\link[=tuv_aq_defaults]{tuv_aq_defaults()}}}

\item{write}{should the options be written to \code{inp_aq} in the TUV directory?
Default \code{TRUE}.}

\item{tuv_dir}{the directory where the compiled TUV executable is located}
}
\value{
the options as a character vector, invisibly
}
\description{
Set required and optional parameters for TUV
}
\examples{
# Setting DOC
set_tuv_aq_params(
 depth_m = 0.25,
 lat = 49.601632,
 lon = -119.605862,
 elev_km = 0.342,
 DOC = 5,
 date = "2023-06-21"
)
# Setting Kd directly (with a different reference wavelength)
set_tuv_aq_params(
 depth_m = 0.25,
 lat = 49.601632,
 lon = -119.605862,
 elev_km = 0.342,
 Kd_ref = 40,
 Kd_wvl = 280,
 date = "2023-06-21"
)

}
\seealso{
\code{\link[=tuv_aq_defaults]{tuv_aq_defaults()}}
}
