% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-tuv.R
\name{tuv}
\alias{tuv}
\title{Call TUV program}
\usage{
tuv(tuv_dir = tuv_data_dir(), quiet = FALSE)
}
\arguments{
\item{tuv_dir}{the directory where the compiled TUV executable is located}

\item{quiet}{Should the progress of the TUV program be printed to the console?}
}
\description{
Call TUV program
}
