% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-tuv.R
\name{view_tuv_aq_params}
\alias{view_tuv_aq_params}
\title{View TUV aquatics options, as set by \code{set_tuv_aq_params()}}
\usage{
view_tuv_aq_params(as_character = FALSE, tuv_dir = tuv_data_dir())
}
\arguments{
\item{as_character}{Return as a character vector? Default \code{FALSE}, in
which case it just prints the parameter list to the screen.}

\item{tuv_dir}{the directory where the compiled TUV executable is located}
}
\description{
View TUV aquatics options, as set by \code{set_tuv_aq_params()}
}
